+++ priv/repo/migrations/20220726115416_alter_remove_type_params_from_rules.exs
@@ -1,0 +1,9 @@
+defmodule TdDd.Repo.Migrations.AlterRemoveTypeParamsFromRules do
+  use Ecto.Migration
+
+  def change do
+    alter table("rules") do
+      remove(:type_params, :map, null: true)
+    end
+  end
+end